feat: switch hello-cosmo-go to WASI preview2 tinygo w/ wash build
Work is underway in wash to enable building components
with the tinygo toolchain. This means that running `wash build` should
be enough for people to get a preview2 tinygo component built and
signed.

However, the support in this repository is *pre* WASI resources
landing, so a slightly older version of some
dependencies (specifically, wasi:io/streams) must be used.

This commit updates the hello-cosmo-go example to use the options that
will work soon (tm) in wash, and the necessary WIT to make the actor
compile properly.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>